+++ shpc/client/config.py
@@ -2,6 +2,7 @@ __author__ = "Vanessa Sochat"
 __copyright__ = "Copyright 2021, Vanessa Sochat"
 __license__ = "MPL 2.0"
 
+import shpc.defaults as defaults
 from shpc.logger import logger
 import sys
 
@@ -18,15 +19,21 @@ def main(args, parser, extra, subparser):
     # The first "param" is either set of get
     command = args.params.pop(0)
 
+    # If the user wants the central config file
+    if args.central:
+        args.settings_file = defaults.default_settings_file
+
     validate = True if not command == "edit" else False
     cli = get_client(
         quiet=args.quiet, settings_file=args.settings_file, validate=validate
     )
 
     # For each new setting, update and save!
+    if command == "inituser":
+        return cli.settings.inituser()
     if command == "edit":
         return cli.settings.edit()
-    elif command == "set":
+    elif command in ["set", "add", "remove"]:
         for param in args.params:
             if ":" not in param:
                 logger.warning(
@@ -35,8 +42,15 @@ def main(args, parser, extra, subparser):
                 )
                 continue
             key, value = param.split(":", 1)
-            cli.settings.set(key, value)
-            logger.info("Updated %s to be %s" % (key, value))
+            if command == "set":
+                cli.settings.set(key, value)
+                logger.info("Updated %s to be %s" % (key, value))
+            elif command == "add":
+                cli.settings.add(key, value)
+                logger.info("Added %s to %s" % (key, value))
+            elif command == "remove":
+                cli.settings.remove(key, value)
+                logger.info("Removed %s from %s" % (key, value))
 
         # Save settings
         cli.settings.save()