+++ shpc/main/settings.py
@@ -7,15 +7,24 @@ from shpc.logger import logger
 import shpc.defaults as defaults
 import shpc.main.schemas
 import shpc.utils
+import shutil
 
 try:
     from ruamel_yaml import YAML
+    from ruamel_yaml.comments import CommentedSeq
 except:
     from ruamel.yaml import YAML
+    from ruamel.yaml.comments import CommentedSeq
 
 from datetime import datetime
 import jsonschema
 import os
+
+
+def OrderedList(*l):
+    ret = CommentedSeq(l)
+    ret.fa.set_flow_style()
+    return ret
 
 
 class SettingsBase:
@@ -39,6 +48,21 @@ class SettingsBase:
         """
         jsonschema.validate(instance=self._settings, schema=shpc.main.schemas.settings)
 
+    def inituser(self):
+        """
+        Create a user specific config in user's home.
+        """
+        user_home = os.path.dirname(defaults.user_settings_file)
+        if not os.path.exists(user_home):
+            os.makedirs(user_home)
+        if os.path.exists(defaults.user_settings_file):
+            logger.exit(
+                "%s already exists! Remove first before re-creating."
+                % defaults.user_settings_file
+            )
+        shutil.copyfile(self.settings_file, defaults.user_settings_file)
+        logger.info("Created user settings file %s" % defaults.user_settings_file)
+
     def edit(self):
         """
         Interactively edit a config file.
@@ -47,11 +71,24 @@ class SettingsBase:
             logger.exit("Settings file not found.")
         shpc.utils.run_command([self.config_editor, self.settings_file], stream=True)
 
+    def get_settings_file(self, settings_file=None):
+        """
+        Get the preferred used settings file.
+        """
+        # Only consider user settings if the file exists!
+        user_settings = None
+        if os.path.exists(defaults.user_settings_file):
+            user_settings = defaults.user_settings_file
+
+        # First preference to command line, then user settings, then default
+        return settings_file or user_settings or defaults.default_settings_file
+
     def load(self, settings_file=None):
         """
         Load the settings file into the settings object
         """
-        self.settings_file = settings_file or defaults.default_settings_file
+        # Get the preferred settings flie
+        self.settings_file = self.get_settings_file(settings_file)
 
         # Exit quickly if the settings file does not exist
         if not os.path.exists(self.settings_file):
@@ -59,20 +96,66 @@ class SettingsBase:
 
         # Default to round trip so we can save comments
         yaml = YAML()
+        yaml.preserve_quotes = True
 
         # Store the original settings for update as we go
         with open(self.settings_file, "r") as fd:
             self._settings = yaml.load(fd.read())
 
     def get(self, key, default=None):
+        """
+        Get a settings value, doing appropriate substitution and expansion.
+        """
         value = self._settings.get(key, default)
-        return self._substitutions(value)
+        value = self._substitutions(value)
+        # If we allow environment substitution, do it
+        if key in defaults.allowed_envars and value:
+            if isinstance(value, list):
+                value = [os.path.expandvars(v) for v in value]
+            else:
+                value = os.path.expandvars(value)
+        return value
 
     def __getattr__(self, key):
         """
         A direct get of an attribute, but default to None if doesn't exist
         """
         return self.get(key)
+
+    def add(self, key, value):
+        """
+        Add a value to a list parameter
+        """
+        # We can only add to lists
+        current = self._settings.get(key)
+        if current and not isinstance(current, list):
+            logger.exit("You cannot only add to a list variable.")
+
+        if value not in current:
+            # Add to the beginning of the list
+            current = [value] + current
+            self._settings[key] = OrderedList()
+            [self._settings[key].append(x) for x in current]
+            self.change_validate(key, value)
+            logger.warning(
+                "Warning: Check with shpc config edit - ordering of list can change."
+            )
+
+    def remove(self, key, value):
+        """
+        Remove a value from a list parameter
+        """
+        current = self._settings.get(key)
+        if current and not isinstance(current, list):
+            logger.exit("You cannot only remove from a list variable.")
+        if not current or value not in current:
+            logger.exit("%s is not in %s" % (value, key))
+        current.pop(current.index(value))
+        self._settings[key] = current
+        self.change_validate(key, current)
+        logger.warning(
+            "Warning: Check with shpc config edit - ordering of list can change."
+        )
 
     def set(self, key, value):
         """
@@ -81,6 +164,11 @@ class SettingsBase:
         value = True if value == "true" else value
         value = False if value == "false" else value
 
+        # List values not allowed for set
+        current = self._settings.get(key)
+        if current and isinstance(current, list):
+            logger.exit("You cannot use 'set' for a list. Use add/remove instead.")
+
         # This is a reference to a dictionary (object) setting
         if ":" in key:
             key, subkey = key.split(":")
@@ -88,6 +176,13 @@ class SettingsBase:
         else:
             self._settings[key] = value
 
+        # Validate and catch error message cleanly
+        self.change_validate(key, value)
+
+    def change_validate(self, key, value):
+        """
+        A courtesy function to validate a new config addition.
+        """
         # Don't allow the user to add a setting not known
         try:
             self.validate()
@@ -120,10 +215,14 @@ class SettingsBase:
             del self._settings[key]
 
     def save(self, filename=None):
+        """
+        Save settings, but do not change order of anything.
+        """
         filename = filename or self.settings_file
         if not filename:
             logger.exit("A filename is required to save to.")
         yaml = YAML()
+
         with open(filename, "w") as fd:
             yaml.dump(self._settings, fd)